Prefer `jest` everywhere instead of using `mocha` for `e2e` tests.
