+++ e2e/__tests__/appium-test-wdio.js.old
@@ -1,5 +1,4 @@
 const wdio = require('webdriverio');
-const assert = require('chai').assert;
 const path = require('path');
 
 
@@ -24,12 +23,14 @@ const androidOptions = Object.assign(
   serverConfig
 );
 
-const app = path.resolve(__dirname, '..', 'android', 'app', 'build', 'outputs', 'apk', 'release', 'app-release.apk')
+const app = path.resolve(__dirname, '../../android/app/build/outputs/apk/release/app-release.apk');
 androidOptions.capabilities.app = app;
 
 let client;
 
-before(async function () {
+beforeAll(async () => {
+  jest.setTimeout(30 * 1000);
+
   client = await wdio.remote(androidOptions);
 
   // NOTE-RT: Setting an implicit wait timeout isn't supported for mobile yet...
@@ -38,82 +39,82 @@ before(async function () {
   // })
 });
 
-after(async function () {
+afterAll(async () => {
   await client.deleteSession();
 });
 
-describe('Create Android session', function () {
-  beforeEach(async function () {
+describe('Create Android session (wdio)', () => {
+  beforeEach(async () => {
     console.info('[beforeEach] client.getPageSource()', await client.getPageSource()); // eslint-disable-line no-console
     console.info('[beforeEach] pausing for 5 seconds between tests...'); // eslint-disable-line no-console
-    await client.pause(5000)
-  });
-
-  it('should create and destroy a session', async function () {
-    assert.strictEqual(await client.getCurrentPackage(), 'com.apktest');
-  });
-
-  it('Press alert button and press OK', async function () {
     await client.pause(5000);
-    console.log('REGULAR BUTTON')
-    let button = await client.$("~test-button-1");
+  });
+
+  it('should create and destroy a session', async () => {
+    expect(await client.getCurrentPackage()).toBe('com.apktest');
+  });
+
+  it('Press alert button and press OK', async () => {
+    await client.pause(5000);
+    console.log('REGULAR BUTTON');
+    let button = await client.$('~test-button-1');
     // console.log(button)
-    button.click()
+    await button.click();
 
     // Click alert message button
     await client.pause(2000);
     let okbutton = await client.$('android=new UiSelector().resourceId("android:id/button1")');
-    console.log('OK')
+    console.log('OK');
     // console.log(okbutton);
-    okbutton.click()
+    await okbutton.click();
     await client.pause(5000);
   });
 
-  it('Press alert button and press Cancel', async function () {
+  it('Press alert button and press Cancel', async () => {
     await client.pause(5000);
-    console.log('REGULAR BUTTON')
-    let button = await client.$("~test-button-1");
+    console.log('REGULAR BUTTON');
+    let button = await client.$('~test-button-1');
     // console.log(button)
-    button.click()
+    await button.click();
 
     // Click alert message button
     await client.pause(2000);
     let okbutton = await client.$('android=new UiSelector().resourceId("android:id/button2")');
-    console.log('CANCEL')
+    console.log('CANCEL');
     // console.log(okbutton);
-    okbutton.click()
+    await okbutton.click();
     await client.pause(5000);
   });
 
-  it('Press second button and press OK', async function () {
+  it('Press second button and press OK', async () => {
     await client.pause(5000);
-    console.log('REGULAR BUTTON')
-    let button = await client.$("~test-button-2");
+    console.log('REGULAR BUTTON');
+    let button = await client.$('~test-button-2');
     // console.log(button)
-    button.click()
+    await button.click();
 
     // Click alert message button
     await client.pause(2000);
     let okbutton = await client.$('android=new UiSelector().resourceId("android:id/button1")');
-    console.log('OK')
+    console.log('OK');
     // console.log(okbutton);
-    okbutton.click()
+    await okbutton.click();
     await client.pause(5000);
   });
 
-  it('Press alert button and press Cancel', async function () {
+  it('Press alert button and press Cancel', async () => {
     await client.pause(5000);
-    console.log('REGULAR BUTTON')
-    let button = await client.$("~test-button-2");
+    console.log('REGULAR BUTTON');
+    let button = await client.$('~test-button-2');
     // console.log(button)
-    button.click()
+    await button.click();
 
     // Click alert message button
     await client.pause(2000);
     let okbutton = await client.$('android=new UiSelector().resourceId("android:id/button2")');
-    console.log('CANCEL')
+    console.log('CANCEL');
     // console.log(okbutton);
-    okbutton.click()
+    await okbutton.click();
     await client.pause(5000);
   });
 });
