+++ e2e/jest.config.json
@@ -1,0 +1,3 @@
+{
+  "verbose": true
+}
